+++ libs/rhino/extraction/ExtractionCore.cs
@@ -12,32 +12,6 @@ namespace Arsenal.Rhino.Extraction;
 
 /// <summary>Internal extraction algorithms with Rhino SDK geometry processing.</summary>
 internal static class ExtractionCore {
-    private static readonly FrozenDictionary<(byte, Type), V> _validationConfig =
-        new Dictionary<(byte, Type), V> {
-            [(1, typeof(GeometryBase))] = V.Standard,
-            [(1, typeof(Brep))] = V.Standard | V.MassProperties,
-            [(1, typeof(Curve))] = V.Standard | V.AreaCentroid,
-            [(1, typeof(Surface))] = V.Standard | V.AreaCentroid,
-            [(1, typeof(Mesh))] = V.Standard | V.MassProperties,
-            [(1, typeof(PointCloud))] = V.Standard,
-            [(2, typeof(GeometryBase))] = V.BoundingBox,
-            [(3, typeof(NurbsCurve))] = V.Standard,
-            [(3, typeof(NurbsSurface))] = V.Standard,
-            [(3, typeof(Curve))] = V.Standard,
-            [(3, typeof(Surface))] = V.Standard,
-            [(4, typeof(Curve))] = V.Standard | V.Degeneracy,
-            [(5, typeof(Curve))] = V.Tolerance,
-            [(6, typeof(Brep))] = V.Standard | V.Topology,
-            [(6, typeof(Mesh))] = V.Standard | V.MeshSpecific,
-            [(6, typeof(Curve))] = V.Standard,
-            [(7, typeof(Brep))] = V.Standard | V.Topology,
-            [(7, typeof(Mesh))] = V.Standard | V.MeshSpecific,
-            [(10, typeof(Curve))] = V.Standard | V.Degeneracy,
-            [(10, typeof(Surface))] = V.Standard,
-            [(11, typeof(Curve))] = V.Standard | V.Degeneracy,
-            [(12, typeof(Curve))] = V.Standard,
-            [(13, typeof(Curve))] = V.Standard,
-        }.ToFrozenDictionary();
 
     [Pure, MethodImpl(MethodImplOptions.AggressiveInlining)]
     internal static Result<IReadOnlyList<Point3d>> Execute(GeometryBase geometry, object spec, IGeometryContext context) {
@@ -65,8 +39,8 @@ internal static class ExtractionCore {
         };
 
         try {
-            V mode = _validationConfig.TryGetValue((kind, normalized.GetType()), out V exact) ? exact :
-                _validationConfig.Where(kv => kv.Key.Item1 == kind && kv.Key.Item2.IsInstanceOfType(normalized))
+            V mode = ExtractionConfig.ValidationModes.TryGetValue((kind, normalized.GetType()), out V exact) ? exact :
+                ExtractionConfig.ValidationModes.Where(kv => kv.Key.Item1 == kind && kv.Key.Item2.IsInstanceOfType(normalized))
                     .OrderByDescending(kv => kv.Key.Item2, Comparer<Type>.Create(static (a, b) => a.IsAssignableFrom(b) ? -1 : b.IsAssignableFrom(a) ? 1 : 0))
                     .Select(kv => kv.Value)
                     .DefaultIfEmpty(V.Standard)
