+++ libs/rhino/topology/TopologyCore.cs
@@ -11,137 +11,7 @@ using Rhino;
 using Rhino.Geometry;
 
 namespace Arsenal.Rhino.Topology;
-
-/// <summary>Edge continuity classification enumeration for geometric analysis.</summary>
-[System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1028:Enum Storage should be Int32", Justification = "byte enum for performance and memory efficiency")]
-[System.Diagnostics.CodeAnalysis.SuppressMessage("Naming", "MA0048:File name must match type name", Justification = "Result types grouped semantically in TopologyCompute.cs")]
-public enum EdgeContinuityType : byte {
-    /// <summary>G0 discontinuous or below minimum continuity threshold.</summary>
-    Sharp = 0,
-    /// <summary>G1 continuous (tangent continuity).</summary>
-    Smooth = 1,
-    /// <summary>G2 continuous (curvature continuity).</summary>
-    Curvature = 2,
-    /// <summary>Interior manifold edge (valence=2, meets continuity requirement).</summary>
-    Interior = 3,
-    /// <summary>Boundary naked edge (valence=1).</summary>
-    Boundary = 4,
-    /// <summary>Non-manifold edge (valence>2).</summary>
-    NonManifold = 5,
-}
-
-/// <summary>Naked edge analysis result containing edge curves and indices with topology metadata.</summary>
-[DebuggerDisplay("{DebuggerDisplay}")]
-[System.Diagnostics.CodeAnalysis.SuppressMessage("Naming", "MA0048:File name must match type name", Justification = "Result records grouped semantically in TopologyCompute.cs")]
-public sealed record NakedEdgeData(
-    IReadOnlyList<Curve> EdgeCurves,
-    IReadOnlyList<int> EdgeIndices,
-    IReadOnlyList<int> Valences,
-    bool IsOrdered,
-    int TotalEdgeCount,
-    double TotalLength) : Topology.IResult {
-    [Pure]
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Style", "IDE0009:Member access should be qualified", Justification = "this. qualification not required in records")]
-    private string DebuggerDisplay => string.Create(
-        CultureInfo.InvariantCulture,
-        $"NakedEdges: {EdgeCurves.Count}/{TotalEdgeCount} | L={TotalLength:F3} | Ordered={IsOrdered}");
-}
-
-/// <summary>Boundary loop analysis result with closed loop curves and join diagnostics.</summary>
-[DebuggerDisplay("{DebuggerDisplay}")]
-[System.Diagnostics.CodeAnalysis.SuppressMessage("Naming", "MA0048:File name must match type name", Justification = "Result records grouped semantically in TopologyCompute.cs")]
-public sealed record BoundaryLoopData(
-    IReadOnlyList<Curve> Loops,
-    IReadOnlyList<IReadOnlyList<int>> EdgeIndicesPerLoop,
-    IReadOnlyList<double> LoopLengths,
-    IReadOnlyList<bool> IsClosedPerLoop,
-    double JoinTolerance,
-    int FailedJoins) : Topology.IResult {
-    [Pure]
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Style", "IDE0009:Member access should be qualified", Justification = "this. qualification not required in records")]
-    private string DebuggerDisplay => string.Create(
-        CultureInfo.InvariantCulture,
-        $"BoundaryLoops: {Loops.Count} | FailedJoins={FailedJoins} | Tol={JoinTolerance:E2}");
-}
-
-/// <summary>Non-manifold topology analysis result with diagnostic data for irregular vertices and edges.</summary>
-[DebuggerDisplay("{DebuggerDisplay}")]
-[System.Diagnostics.CodeAnalysis.SuppressMessage("Naming", "MA0048:File name must match type name", Justification = "Result records grouped semantically in TopologyCompute.cs")]
-public sealed record NonManifoldData(
-    IReadOnlyList<int> EdgeIndices,
-    IReadOnlyList<int> VertexIndices,
-    IReadOnlyList<int> Valences,
-    IReadOnlyList<Point3d> Locations,
-    bool IsManifold,
-    bool IsOrientable,
-    int MaxValence) : Topology.IResult {
-    [Pure]
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Style", "IDE0009:Member access should be qualified", Justification = "this. qualification not required in records")]
-    private string DebuggerDisplay => IsManifold
-        ? "Manifold: No issues detected"
-        : string.Create(
-            CultureInfo.InvariantCulture,
-            $"NonManifold: Edges={EdgeIndices.Count} | Verts={VertexIndices.Count} | MaxVal={MaxValence}");
-}
-
-/// <summary>Connected component analysis result with adjacency graph and spatial bounds.</summary>
-[DebuggerDisplay("{DebuggerDisplay}")]
-[System.Diagnostics.CodeAnalysis.SuppressMessage("Naming", "MA0048:File name must match type name", Justification = "Result records grouped semantically in TopologyCompute.cs")]
-public sealed record ConnectivityData(
-    IReadOnlyList<IReadOnlyList<int>> ComponentIndices,
-    IReadOnlyList<int> ComponentSizes,
-    IReadOnlyList<BoundingBox> ComponentBounds,
-    int TotalComponents,
-    bool IsFullyConnected,
-    FrozenDictionary<int, IReadOnlyList<int>> AdjacencyGraph) : Topology.IResult {
-    [Pure]
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Style", "IDE0009:Member access should be qualified", Justification = "this. qualification not required in records")]
-    private string DebuggerDisplay => IsFullyConnected
-        ? "Connectivity: Single connected component"
-        : string.Create(
-            CultureInfo.InvariantCulture,
-            $"Connectivity: {TotalComponents} components | Largest={ComponentSizes.Max()}");
-}
-
-/// <summary>Edge classification result by continuity type with geometric measures.</summary>
-[DebuggerDisplay("{DebuggerDisplay}")]
-[System.Diagnostics.CodeAnalysis.SuppressMessage("Naming", "MA0048:File name must match type name", Justification = "Result records grouped semantically in TopologyCompute.cs")]
-public sealed record EdgeClassificationData(
-    IReadOnlyList<int> EdgeIndices,
-    IReadOnlyList<EdgeContinuityType> Classifications,
-    IReadOnlyList<double> ContinuityMeasures,
-    FrozenDictionary<EdgeContinuityType, IReadOnlyList<int>> GroupedByType,
-    Continuity MinimumContinuity) : Topology.IResult {
-    [Pure]
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Style", "IDE0009:Member access should be qualified", Justification = "this. qualification not required in records")]
-    private string DebuggerDisplay => string.Create(
-        CultureInfo.InvariantCulture,
-        $"EdgeClassification: Total={EdgeIndices.Count} | Sharp={GroupedByType.GetValueOrDefault(EdgeContinuityType.Sharp, []).Count}");
-}
-
-/// <summary>Face adjacency query result with neighbor data and dihedral angle computation.</summary>
-[DebuggerDisplay("{DebuggerDisplay}")]
-[System.Diagnostics.CodeAnalysis.SuppressMessage("Naming", "MA0048:File name must match type name", Justification = "Result records grouped semantically in TopologyCompute.cs")]
-public sealed record AdjacencyData(
-    int EdgeIndex,
-    IReadOnlyList<int> AdjacentFaceIndices,
-    IReadOnlyList<Vector3d> FaceNormals,
-    double DihedralAngle,
-    bool IsManifold,
-    bool IsBoundary) : Topology.IResult {
-    [Pure]
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Style", "IDE0009:Member access should be qualified", Justification = "this. qualification not required in records")]
-    private string DebuggerDisplay => IsBoundary
-        ? string.Create(CultureInfo.InvariantCulture, $"Edge[{EdgeIndex}]: Boundary (valence=1)")
-        : IsManifold
-            ? string.Create(
-                CultureInfo.InvariantCulture,
-                $"Edge[{EdgeIndex}]: Manifold | Angle={DihedralAngle * 180.0 / Math.PI:F1}°")
-            : string.Create(CultureInfo.InvariantCulture, $"Edge[{EdgeIndex}]: NonManifold (valence={AdjacentFaceIndices.Count})");
-}
-
-/// <summary>Internal topology computation algorithms with FrozenDictionary-based type dispatch.</summary>
-internal static class TopologyCompute {
+internal static class TopologyCore {
     private static readonly IReadOnlyList<int> EmptyIndices = [];
 
     [Pure]
@@ -537,14 +407,14 @@ internal static class TopologyCompute {
     [Pure]
     private static Result<IReadOnlyList<EdgeClassificationData>> ExecuteBrepEdgeClassification(Brep brep, Continuity minContinuity, double angleThreshold) {
         IReadOnlyList<int> edgeIndices = [.. Enumerable.Range(0, brep.Edges.Count),];
-        IReadOnlyList<EdgeContinuityType> classifications = [.. edgeIndices.Select(i => ClassifyBrepEdge(
+        IReadOnlyList<EdgeContinuity> classifications = [.. edgeIndices.Select(i => ClassifyBrepEdge(
             edge: brep.Edges[i],
             _: brep,
             minContinuity: minContinuity,
             angleThreshold: angleThreshold)),
         ];
         IReadOnlyList<double> measures = [.. edgeIndices.Select(i => brep.Edges[i].GetLength()),];
-        FrozenDictionary<EdgeContinuityType, IReadOnlyList<int>> grouped = edgeIndices
+        FrozenDictionary<EdgeContinuity, IReadOnlyList<int>> grouped = edgeIndices
             .Select((idx, pos) => (idx, type: classifications[pos]))
             .GroupBy(x => x.type, x => x.idx)
             .ToFrozenDictionary(g => g.Key, g => (IReadOnlyList<int>)[.. g,]);
@@ -559,34 +429,34 @@ internal static class TopologyCompute {
     }
 
     [Pure]
-    private static EdgeContinuityType ClassifyBrepEdge(BrepEdge edge, Brep _, Continuity minContinuity, double angleThreshold) =>
+    private static EdgeContinuity ClassifyBrepEdge(BrepEdge edge, Brep _, Continuity minContinuity, double angleThreshold) =>
         edge.Valence switch {
-            EdgeAdjacency.Naked => EdgeContinuityType.Boundary,
-            EdgeAdjacency.NonManifold => EdgeContinuityType.NonManifold,
+            EdgeAdjacency.Naked => EdgeContinuity.Boundary,
+            EdgeAdjacency.NonManifold => EdgeContinuity.NonManifold,
             EdgeAdjacency.Interior => edge.EdgeCurve switch {
-                Curve crv when crv.IsContinuous(continuityType: Continuity.G2_continuous, t: crv.Domain.Mid) || crv.IsContinuous(continuityType: Continuity.G2_locus_continuous, t: crv.Domain.Mid) => EdgeContinuityType.Curvature,
-                Curve crv when edge.IsSmoothManifoldEdge(angleToleranceRadians: angleThreshold) || crv.IsContinuous(continuityType: Continuity.G1_continuous, t: crv.Domain.Mid) || crv.IsContinuous(continuityType: Continuity.G1_locus_continuous, t: crv.Domain.Mid) => EdgeContinuityType.Smooth,
-                _ when minContinuity >= Continuity.G1_continuous => EdgeContinuityType.Sharp,
-                _ => EdgeContinuityType.Interior,
+                Curve crv when crv.IsContinuous(continuityType: Continuity.G2_continuous, t: crv.Domain.Mid) || crv.IsContinuous(continuityType: Continuity.G2_locus_continuous, t: crv.Domain.Mid) => EdgeContinuity.Curvature,
+                Curve crv when edge.IsSmoothManifoldEdge(angleToleranceRadians: angleThreshold) || crv.IsContinuous(continuityType: Continuity.G1_continuous, t: crv.Domain.Mid) || crv.IsContinuous(continuityType: Continuity.G1_locus_continuous, t: crv.Domain.Mid) => EdgeContinuity.Smooth,
+                _ when minContinuity >= Continuity.G1_continuous => EdgeContinuity.Sharp,
+                _ => EdgeContinuity.Interior,
             },
-            _ => EdgeContinuityType.Sharp,
+            _ => EdgeContinuity.Sharp,
         };
 
     [Pure]
     private static Result<IReadOnlyList<EdgeClassificationData>> ExecuteMeshEdgeClassification(Mesh mesh, double angleThreshold) {
         double curvatureThreshold = angleThreshold * 0.1;
         IReadOnlyList<int> edgeIndices = [.. Enumerable.Range(0, mesh.TopologyEdges.Count),];
-        IReadOnlyList<EdgeContinuityType> classifications = [.. edgeIndices.Select(i => {
+        IReadOnlyList<EdgeContinuity> classifications = [.. edgeIndices.Select(i => {
             int[] connectedFaces = mesh.TopologyEdges.GetConnectedFaces(i);
             return connectedFaces.Length switch {
-                1 => EdgeContinuityType.Boundary,
-                > 2 => EdgeContinuityType.NonManifold,
+                1 => EdgeContinuity.Boundary,
+                > 2 => EdgeContinuity.NonManifold,
                 2 => CalculateDihedralAngle(mesh: mesh, faceIdx1: connectedFaces[0], faceIdx2: connectedFaces[1]) switch {
-                    double angle when Math.Abs(angle) < curvatureThreshold => EdgeContinuityType.Curvature,
-                    double angle when Math.Abs(angle) < angleThreshold => EdgeContinuityType.Smooth,
-                    _ => EdgeContinuityType.Sharp,
+                    double angle when Math.Abs(angle) < curvatureThreshold => EdgeContinuity.Curvature,
+                    double angle when Math.Abs(angle) < angleThreshold => EdgeContinuity.Smooth,
+                    _ => EdgeContinuity.Sharp,
                 },
-                _ => EdgeContinuityType.Sharp,
+                _ => EdgeContinuity.Sharp,
             };
         }),
         ];
@@ -595,7 +465,7 @@ internal static class TopologyCompute {
             return mesh.TopologyVertices[verts.I].DistanceTo(mesh.TopologyVertices[verts.J]);
         }),
         ];
-        FrozenDictionary<EdgeContinuityType, IReadOnlyList<int>> grouped = edgeIndices
+        FrozenDictionary<EdgeContinuity, IReadOnlyList<int>> grouped = edgeIndices
             .Select((idx, pos) => (idx, type: classifications[pos]))
             .GroupBy(x => x.type, x => x.idx)
             .ToFrozenDictionary(g => g.Key, g => (IReadOnlyList<int>)[.. g,]);
